+++ src/utils/data/applicationColumn.ts
@@ -1,4 +1,5 @@
 enum APPLICATION_COLUMN {
+  SERIAL_NUMBER = 'SERIAL_NUMBER',
   LAST_ACTIVE_DATE = 'LAST_ACTIVE_DATE',
   DEADLINE_DATE = 'DEADLINE_DATE',
   APPLICATION_NAME = 'APPLICATION_NAME',
