+++ src/utils/data/columnsPerUserRole.ts
@@ -1,8 +1,9 @@
-import { ColumnsPerRole } from '../../utils/types'
+import { ColumnsPerRole } from '../types'
 import APPLICATION_COLUMN from './applicationColumn'
 
 const COLUMNS_PER_ROLE: ColumnsPerRole = {
   applicant: [
+    APPLICATION_COLUMN.SERIAL_NUMBER,
     APPLICATION_COLUMN.LAST_ACTIVE_DATE,
     APPLICATION_COLUMN.DEADLINE_DATE,
     APPLICATION_COLUMN.APPLICATION_NAME,
@@ -11,6 +12,7 @@ const COLUMNS_PER_ROLE: ColumnsPerRole = {
     APPLICATION_COLUMN.OUTCOME,
   ],
   reviewer1: [
+    APPLICATION_COLUMN.SERIAL_NUMBER,
     APPLICATION_COLUMN.LAST_ACTIVE_DATE,
     APPLICATION_COLUMN.APPLICATION_NAME,
     APPLICATION_COLUMN.APPLICANT,
@@ -21,6 +23,7 @@ const COLUMNS_PER_ROLE: ColumnsPerRole = {
     APPLICATION_COLUMN.OUTCOME,
   ],
   reviewer2: [
+    APPLICATION_COLUMN.SERIAL_NUMBER,
     APPLICATION_COLUMN.LAST_ACTIVE_DATE,
     APPLICATION_COLUMN.APPLICATION_NAME,
     APPLICATION_COLUMN.APPLICANT,
@@ -31,6 +34,7 @@ const COLUMNS_PER_ROLE: ColumnsPerRole = {
     APPLICATION_COLUMN.OUTCOME,
   ],
   supervisor: [
+    APPLICATION_COLUMN.SERIAL_NUMBER,
     APPLICATION_COLUMN.LAST_ACTIVE_DATE,
     APPLICATION_COLUMN.APPLICATION_NAME,
     APPLICATION_COLUMN.APPLICANT,
@@ -39,6 +43,7 @@ const COLUMNS_PER_ROLE: ColumnsPerRole = {
     APPLICATION_COLUMN.OUTCOME,
   ],
   consolidator: [
+    APPLICATION_COLUMN.SERIAL_NUMBER,
     APPLICATION_COLUMN.LAST_ACTIVE_DATE,
     APPLICATION_COLUMN.APPLICATION_NAME,
     APPLICATION_COLUMN.APPLICANT,
