--- src/utils/model/index.ts
@@ -1,4 +1,0 @@
-import APPLICATION_COLUMN from './applicationColumn'
-import USER_ROLE from './userRole'
-
-export { APPLICATION_COLUMN, USER_ROLE }
